Optimize audio engine and harden numeric precision hygiene
This commit implements several enhancements to the Mixxx audio engine,
focusing on real-time performance and numeric precision hygiene:

1.  **SampleUtil Optimization**: Hoisted the `monoMixdown` conditional
    out of the hot loop in `finalizeMainMix`. This allows the compiler
    to generate distinct, highly-vectorizable paths for stereo and mono
    modes. Applied `M_RESTRICT` to input/output buffers to assist SIMD
    auto-vectorization by guaranteeing no memory aliasing.

2.  **Numeric Precision Hygiene**: Audited the engine's critical paths for
    unintentional `double` promotions. Updated calls to `std::log10`,
    `std::pow`, and `std::exp` in `EngineVuMeter`, `EnginePregain`,
    `EngineXfader`, and `EngineFilterMoogLadder4` to ensure `float`
    overloads are used consistently.

3.  **EngineVuMeter Refactoring**: Clarified the peak indicator holding
    logic. Replaced confusing magic numbers with an explicit calculation
    based on the desired 500ms duration, sample rate, and buffer size.

4.  **Math Utility Modernization**: Removed `using std::fabs` from
    `src/util/math.h` to avoid global namespace pollution and promote the
    use of type-safe `std::abs` overloads.

Verified with `mixxx-test` suite (EngineBuffer, EngineMixer,
EngineFilterBiquad, and SampleUtil tests passed).

+++ src/engine/enginexfader.cpp
@@ -51,13 +51,13 @@ void EngineXfader::getXfadeGains(double xfadePosition,
     } else {
         if (xfadePositionLeft < 0) { // on left side
             xfadePositionLeft *= -1;
-            *gain2 = 1.0f - static_cast<float>(std::pow(xfadePositionLeft, transform));
+            *gain2 = 1.0f - static_cast<float>(std::pow(static_cast<float>(xfadePositionLeft), static_cast<float>(transform)));
         } else {
             *gain2 = 1.0f;
         }
 
         if (xfadePositionRight > 0) { // right side
-            *gain1 = 1.0f - static_cast<float>(std::pow(xfadePositionRight, transform));
+            *gain1 = 1.0f - static_cast<float>(std::pow(static_cast<float>(xfadePositionRight), static_cast<float>(transform)));
         } else {
             *gain1 = 1.0f;
         }
